Coinjoin payment batching
This PR introduces support to the internals to allow payment batching. The destination provider supplies a list of `PendingPayment`. The `IWallet` has a new bool that signals its intent to batch payments to the coinjoin client. Please note that for this to  be efficient in payments, wasabi wallet needs to:
* not stop the coinjoin client when all mixed if payments are enabled
* the coin selector needs to factor in any payments it wants to do
* the coordinator needs to remove its output script type limitations